app-generation/app-artefacts/engenharia/backend/rf002-multimodal-ingestion/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Celery and Redis
celery==5.3.6
redis==5.0.1

# File processing
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# Audio/Video
openai-whisper==20231117
ffmpeg-python==0.2.0

# Web scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Archive handling
py7zr==0.21.0

# MIME type detection
python-magic==0.4.27

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Observability
structlog==24.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
